refactor(ble): deduplicate handler-noble and handler-noble-legacy (#181)
Extract the ~600 lines duplicated between handler-noble.ts (@stoprocent/noble)
and handler-noble-legacy.ts (@abandonware/noble) into a single parameterized
module handler-noble-shared.ts. A createNobleHandler() factory takes the Noble
instance plus a getState accessor and returns scanAndReadRaw, scanAndRead,
scanDevices and _internals; the two driver files become thin entrypoints.

The state read is the one non-interchangeable difference and is injected per
driver: @stoprocent/noble reads .state (its getter lazily inits bindings, which
is intended), while @abandonware/noble reads the raw ._state field to avoid the
init side effect of its own .state getter. A naive merged read would have fired
that side effect on the legacy driver.

Handler selection precedence, public API and the dynamic-import dead-code
elimination are unchanged. Existing per-driver tests stay green and a new seam
test pins the getState injection.
